invalidate custom translations on sync
