+++ client/packages/host/src/components/Sync/SyncModal.tsx
@@ -10,7 +10,6 @@ import {
   RadioIcon,
   Typography,
   UNDEFINED_STRING_VALUE,
-  CUSTOM_TRANSLATIONS_NAMESPACE,
   useAppTheme,
   useAuthContext,
   useFormatDateTime,
@@ -18,7 +17,7 @@ import {
   useQueryClient,
   useTranslation,
   useMediaQuery,
-  useIntl,
+  useIntlUtils,
 } from '@openmsupply-client/common';
 import { useSync } from '@openmsupply-client/system';
 import { SyncProgress } from '../SyncProgress';
@@ -46,7 +45,7 @@ const useHostSync = (enabled: boolean) => {
   // true by default to wait for first syncStatus api result
   const [isLoading, setIsLoading] = useState(true);
   const queryClient = useQueryClient();
-  const { i18n } = useIntl();
+  const { invalidateCustomTranslations } = useIntlUtils();
 
   useEffect(() => {
     if (!syncStatus) {
@@ -63,7 +62,7 @@ const useHostSync = (enabled: boolean) => {
       allowSleep();
       queryClient.invalidateQueries(); // refresh the page user is on after sync finishes
       // Reload custom translations, in case we received new ones via sync
-      i18n.reloadResources(undefined, CUSTOM_TRANSLATIONS_NAMESPACE);
+      invalidateCustomTranslations();
     }
   }, [syncStatus?.isSyncing]);
